+++ CHANGELOG.md
@@ -17,6 +17,8 @@ list the bigger ones.
   - A `--no-slave-console` switch to disable creation of new console windows
     for slaves in coralslaveprovider. This only has an effect on Windows.
   - User-settable log level via the `--log-level` switch.
+  - File logging, controlled with the `--log-file` and `--log-file-dir`
+    switches.
 ### Changed
   - The slave provider command line interface is now a bit more well-defined.